chore: address issue #613
